feat: Implement developer login and deploy to Firebase
