+++ src/components/AuthPage.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { useAuth } from '../contexts/AuthContext';
-import { BookHeart, Code } from 'lucide-react';
+import { BookHeart } from 'lucide-react';
+import DeveloperLogin from './DeveloperLogin';
 
 export default function AuthPage() {
   const [isLogin, setIsLogin] = useState(true);
@@ -9,11 +10,7 @@ export default function AuthPage() {
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(false);
   const { signIn, signUp, signInWithGoogle } = useAuth();
-
-  const [devFormEmail, setDevFormEmail] = useState('');
-  const [devFormPassword, setDevFormPassword] = useState('');
-  const [devFormPhone, setDevFormPhone] = useState('');
-  const showDevLogin = import.meta.env.DEV;
+  const showDevLogin = true;
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -52,31 +49,6 @@ export default function AuthPage() {
     }
   };
 
-  const handleDevLogin = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError('');
-    setLoading(true);
-
-    try {
-      const { error: authError } = await signIn(devFormEmail, devFormPassword);
-
-      if (authError) {
-        if (authError.message.includes('Invalid login credentials')) {
-          const { error: signUpError } = await signUp(devFormEmail, devFormPassword);
-          if (signUpError) {
-            setError(signUpError.message);
-          }
-        } else {
-          setError(authError.message);
-        }
-      }
-    } catch (err) {
-      setError('An unexpected error occurred');
-    } finally {
-      setLoading(false);
-    }
-  };
-
   return (
     <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-slate-50 to-slate-100 dark:from-slate-900 dark:to-slate-800 px-4">
       <div className="max-w-md w-full">
@@ -94,77 +66,7 @@ export default function AuthPage() {
 
         <div className="bg-white dark:bg-slate-800 rounded-2xl shadow-xl p-8">
           {showDevLogin && (
-            <div className="mb-6 p-4 bg-amber-50 dark:bg-amber-900/20 border-2 border-amber-200 dark:border-amber-800 rounded-lg">
-              <div className="flex items-center gap-2 mb-3">
-                <Code className="text-amber-600 dark:text-amber-400" size={20} />
-                <span className="font-semibold text-amber-900 dark:text-amber-100">Developer Login</span>
-              </div>
-              <form onSubmit={handleDevLogin} className="space-y-3">
-                <div>
-                  <label htmlFor="dev-email" className="block text-xs font-medium text-amber-900 dark:text-amber-100 mb-1">
-                    Email
-                  </label>
-                  <input
-                    id="dev-email"
-                    type="email"
-                    value={devFormEmail}
-                    onChange={(e) => setDevFormEmail(e.target.value)}
-                    required
-                    className="w-full px-3 py-2 rounded-lg border border-amber-300 dark:border-amber-700 bg-white dark:bg-slate-700 text-slate-900 dark:text-white focus:ring-2 focus:ring-amber-500 focus:border-transparent text-sm"
-                    placeholder="dev@example.com"
-                  />
-                </div>
-                <div>
-                  <label htmlFor="dev-password" className="block text-xs font-medium text-amber-900 dark:text-amber-100 mb-1">
-                    Password
-                  </label>
-                  <input
-                    id="dev-password"
-                    type="password"
-                    value={devFormPassword}
-                    onChange={(e) => setDevFormPassword(e.target.value)}
-                    required
-                    minLength={6}
-                    className="w-full px-3 py-2 rounded-lg border border-amber-300 dark:border-amber-700 bg-white dark:bg-slate-700 text-slate-900 dark:text-white focus:ring-2 focus:ring-amber-500 focus:border-transparent text-sm"
-                    placeholder="••••••••"
-                  />
-                </div>
-                <div>
-                  <label htmlFor="dev-phone" className="block text-xs font-medium text-amber-900 dark:text-amber-100 mb-1">
-                    Phone Number (UK)
-                  </label>
-                  <div className="flex gap-2">
-                    <input
-                      type="text"
-                      value="+44"
-                      disabled
-                      className="w-16 px-3 py-2 rounded-lg border border-amber-300 dark:border-amber-700 bg-slate-100 dark:bg-slate-600 text-slate-900 dark:text-white text-sm text-center"
-                      aria-label="Country code"
-                    />
-                    <input
-                      id="dev-phone"
-                      type="tel"
-                      value={devFormPhone}
-                      onChange={(e) => setDevFormPhone(e.target.value.replace(/\D/g, ''))}
-                      required
-                      pattern="[0-9]{10}"
-                      className="flex-1 px-3 py-2 rounded-lg border border-amber-300 dark:border-amber-700 bg-white dark:bg-slate-700 text-slate-900 dark:text-white focus:ring-2 focus:ring-amber-500 focus:border-transparent text-sm"
-                      placeholder="7700900000"
-                      maxLength={10}
-                    />
-                  </div>
-                  <p className="text-xs text-amber-700 dark:text-amber-300 mt-1">Enter 10 digits after +44</p>
-                </div>
-                <button
-                  type="submit"
-                  disabled={loading}
-                  className="w-full bg-amber-500 hover:bg-amber-600 disabled:bg-amber-300 text-white font-medium py-2 px-4 rounded-lg transition-colors flex items-center justify-center gap-2"
-                >
-                  <Code size={18} />
-                  {loading ? 'Logging in...' : 'Dev Login'}
-                </button>
-              </form>
-            </div>
+            <DeveloperLogin />
           )}
 
           <div className="flex gap-2 mb-6" role="group" aria-label="Authentication mode">
